BUG: Fix View3D volume rendering
Volume Actor was shared between 2D and 3D view. A consequence is that the volume rendering shared the same RGB color transfer function and the same scalar opacity transfer function. 2D rendering was not the one expected when changing opacity transfer function.
The solution used here is to define a different actor for the 3D view.